Improved the image quality when viewing images with transparency.
After applying a background color, and with gamma correction enabled, the
colors were corrected in a way that was fast, but could sometimes be pretty
inaccurate. In particular, fully transparent pixels sometimes weren't made
exactly the same color as the background.

After this change, colors are more accurate, but partially-transparent
pixels are processed much more slowly than they were.